+++ R/03_metrics.R
@@ -113,7 +113,7 @@ info_criteria <- function(object, metrics = "all", metadata = TRUE, digits = 2) 
 #'     y = Canopy,
 #'     grp = Plot,
 #'     fn = "fn_logistic",
-#'     parameters = c(L = 100, k = 4, t0 = 50),
+#'     parameters = c(a = 0.199, t0 = 47.7, k = 100),
 #'     subset = 40
 #'   )
 #' print(mod_2)
@@ -187,7 +187,7 @@ performance <- function(..., metrics = "all", metadata = FALSE, digits = 2) {
 #'     y = Canopy,
 #'     grp = Plot,
 #'     fn = "fn_logistic",
-#'     parameters = c(L = 100, k = 4, t0 = 50),
+#'     parameters = c(a = 0.199, t0 = 47.7, k = 100),
 #'     subset = 40
 #'   )
 #' # Model 3