+++ R/99_growth_curves.R
@@ -61,9 +61,9 @@ fn_quad <- function(t, a, b, c) {
 #' Computes a value based on a logistic function.
 #'
 #' @param t Numeric value.
-#' @param L Numeric value.
-#' @param k Numeric value.
+#' @param a Numeric value.
 #' @param t0 Numeric value.
+#' @param k Numeric value.
 #'
 #' @return A numeric value based on the logistic function.
 #' @export
@@ -71,7 +71,7 @@ fn_quad <- function(t, a, b, c) {
 #' @details
 #' \if{html}{
 #' \deqn{
-#' f(t; L, k, t0) = \frac{L}{1 + e^{-k(t - t_0)}}
+#' f(t; a, t0, k) = \frac{k}{1 + e^{-a(t - t_0)}}
 #' }
 #' }
 #'
@@ -79,16 +79,15 @@ fn_quad <- function(t, a, b, c) {
 #' library(flexFitR)
 #' plot_fn(
 #'   fn = "fn_logistic",
-#'   params = c(L = 100, k = 0.199, t0 = 47.7),
+#'   params = c(a = 0.199, t0 = 47.7, k = 100),
 #'   interval = c(0, 108),
 #'   n_points = 2000
 #' )
-fn_logistic <- function(t, L, k, t0) {
-  y <- L / (1 + exp(-k * (t - t0)))
-  return(y)
+fn_logistic <- function(t, a, t0, k) {
+  k / (1 + exp(-a * (t - t0)))
 }
 
-#' Exponential linear function 1
+#' Exponential-linear function
 #'
 #' Computes a value based on an exponential growth curve and linear decay model for time.
 #'
@@ -122,13 +121,13 @@ fn_logistic <- function(t, L, k, t0) {
 #' @examples
 #' library(flexFitR)
 #' plot_fn(
-#'   fn = "fn_exp1_lin",
+#'   fn = "fn_exp_lin",
 #'   params = c(t1 = 35, t2 = 55, alpha = 1 / 20, beta = -1 / 40),
 #'   interval = c(0, 108),
 #'   n_points = 2000,
 #'   auc_label_size = 3
 #' )
-fn_exp1_lin <- function(t, t1, t2, alpha, beta) {
+fn_exp_lin <- function(t, t1, t2, alpha, beta) {
   if (t < t1) {
     return(0)
   }
@@ -141,9 +140,9 @@ fn_exp1_lin <- function(t, t1, t2, alpha, beta) {
   }
 }
 
-#' Exponential linear function 2
+#' Super-exponential linear function
 #'
-#' Computes a value based on an exponential growth curve and linear decay model for time.
+#' Computes a value based on a super exponential growth curve and linear decay model for time.
 #'
 #' @param t Numeric. The time value.
 #' @param t1 Numeric. The lower threshold time. Assumed to be known.
@@ -152,7 +151,7 @@ fn_exp1_lin <- function(t, t1, t2, alpha, beta) {
 #' than 0.
 #' @param beta Numeric. The parameter for the linear term. Must be less than 0.
 #'
-#' @return A numeric value based on the exponential linear model.
+#' @return A numeric value based on the super exponential linear model.
 #' If \code{t} is less than \code{t1}, the function returns 0.
 #' If \code{t} is between \code{t1} and \code{t2} (inclusive),
 #' the function returns \code{exp(alpha * (t - t1)^2) - 1}.
@@ -194,7 +193,7 @@ fn_exp2_lin <- function(t, t1, t2, alpha, beta) {
   }
 }
 
-#' Exponential exponential function 1
+#' Double-exponential function
 #'
 #' Computes a value based on an exponential growth curve and exponential decay model for time.
 #'
@@ -229,14 +228,14 @@ fn_exp2_lin <- function(t, t1, t2, alpha, beta) {
 #' @examples
 #' library(flexFitR)
 #' plot_fn(
-#'   fn = "fn_exp1_exp",
+#'   fn = "fn_exp_exp",
 #'   params = c(t1 = 35, t2 = 55, alpha = 1 / 20, beta = -1 / 30),
 #'   interval = c(0, 108),
 #'   n_points = 2000,
 #'   auc_label_size = 3,
 #'   y_auc_label = 0.2
 #' )
-fn_exp1_exp <- function(t, t1, t2, alpha, beta) {
+fn_exp_exp <- function(t, t1, t2, alpha, beta) {
   if (t < t1) {
     return(0)
   }
@@ -249,9 +248,9 @@ fn_exp1_exp <- function(t, t1, t2, alpha, beta) {
   }
 }
 
-#' Exponential exponential Function 2
+#' Super-exponential exponential function
 #'
-#' Computes a value based on an exponential growth curve and exponential decay model for time.
+#' Computes a value based on a super exponential growth curve and exponential decay model for time.
 #'
 #' @param t Numeric. The time value.
 #' @param t1 Numeric. The lower threshold time. Assumed to be known.
@@ -261,7 +260,7 @@ fn_exp1_exp <- function(t, t1, t2, alpha, beta) {
 #' @param beta Numeric. The parameter for the second exponential term.
 #' Must be less than 0.
 #'
-#' @return A numeric value based on the double exponential model.
+#' @return A numeric value based on the super double exponential model.
 #' If \code{t} is less than \code{t1}, the function returns 0.
 #' If \code{t} is between \code{t1} and \code{t2} (inclusive),
 #' the function returns \code{exp(alpha * (t - t1)^2) - 1}.
@@ -351,35 +350,99 @@ fn_lin_plat <- function(t, t1 = 45, t2 = 80, k = 0.9) {
   return(y)
 }
 
-#' Sum of Squares Error Function for Piecewise Model
+#' Linear-logistic function
 #'
-#' Calculates the sum of squared errors (SSE) between observed values and values
-#' predicted by the \link{fn_lin_plat} function. This is the objective function to
-#' be minimized in the optimx package.
+#' Computes a value based on a linear-logistic growth curve.
 #'
-#' @param params Numeric vector. The parameters for the \code{fn_lin_plat} function,
-#' where \code{params[1]} is \code{t1} and \code{params[2]} is \code{t2}.
-#' @param t Numeric vector. The time values.
-#' @param y Numeric vector. The observed values.
+#' @param t Numeric. The time value.
+#' @param t1 Numeric. The lower threshold time. Default is 45.
+#' @param t2 Numeric. The upper threshold time. Default is 80.
+#' @param k Numeric. The maximum value of the function. Default is 0.9. Assumed to be known.
+#' @return A numeric value based on the linear-logistic growth curve.
+#' If \code{t} is less than \code{t1}, the function returns 0.
+#' If \code{t} is between \code{t1} and \code{t2} (inclusive),
+#' the function returns a value between 0 and \code{k} in a linear trend.
+#' If \code{t} is greater than \code{t2}, the function returns a value based on a logistic curve.
+#' @export
 #'
-#' @return A numeric value representing the sum of squared errors.
-#' @noRd
+#' @details
+#' \if{html}{
+#' \deqn{
+#' f(t; t_1, t_2, k) =
+#' \begin{cases}
+#' 0 & \text{if } t < t_1 \\
+#' \dfrac{k}{2(t_2 - t_1)} \cdot (t - t_1) & \text{if } t_1 \leq t \leq t_2 \\
+#' \dfrac{k}{1 + e^{-2(t - t_2) / (t_2 - t_1)}} & \text{if } t > t_2
+#' \end{cases}
+#' }
+#' }
 #'
 #' @examples
 #' library(flexFitR)
-#' x <- c(0, 29, 36, 42, 56, 76, 92, 100, 108)
-#' y <- c(0, 0, 4.379, 26.138, 78.593, 100, 100, 100, 100)
-#' sse_piwise(params = c(34.9, 61.8), t = x, y = y)
+#' plot_fn(
+#'   fn = "fn_lin_logis",
+#'   params = c(t1 = 35, t2 = 50, k = 100),
+#'   interval = c(0, 108),
+#'   n_points = 2000,
+#'   auc_label_size = 3
+#' )
+fn_lin_logis <- function(t, t1, t2, k) {
+  if (t < t1) {
+    return(0)
+  }
+  if (t > t1 && t < t2) {
+    return(k / 2 / (t2 - t1) * (t - t1))
+  }
+  return(k / (1 + exp(-2 * (t - t2) / (t2 - t1))))
+}
+
+#' Quadratic-plateau function
 #'
-#' y_hat <- sapply(x, FUN = fn_lin_plat, t1 = 34.9, t2 = 61.8, k = 100)
-#' sum((y - y_hat)^2)
-sse_piwise <- function(params, t, y) {
-  t1 <- params[1]
-  t2 <- params[2]
-  k <- max(y)
-  y_hat <- sapply(t, FUN = fn_lin_plat, t1 = t1, t2 = t2, k = k)
-  sse <- sum((y - y_hat)^2)
-  return(sse)
+#' Computes a value based on a quadratic-plateau growth curve.
+#'
+#' @param t Numeric. The time value.
+#' @param t1 Numeric. The lower threshold time. Default is 45.
+#' @param t2 Numeric. The upper threshold time. Default is 80.
+#' @param b Numeric.
+#' @param k Numeric. The maximum value of the function. Default is 0.9. Assumed to be known.
+#' @return A numeric value based on quadratic-plateau growth curve.
+#' If \code{t} is less than \code{t1}, the function returns 0.
+#' If \code{t} is between \code{t1} and \code{t2} (inclusive),
+#' the function returns a value between 0 and \code{k} in a linear trend.
+#' If \code{t} is greater than \code{t2}, the function returns a value based on a logistic curve.
+#' @export
+#'
+#' @details
+#' \if{html}{
+#' \deqn{
+#' f(t; t_1, t_2, b, k) =
+#' \begin{cases}
+#' 0 & \text{if } t < t_1 \\
+#' b (t - t_1) + \frac{k - b (t_2 - t_1)}{(t_2 - t_1)^2} (t - t_1)^2 & \text{if } t_1 \leq t \leq t_2 \\
+#' k & \text{if } t > t_2
+#' \end{cases}
+#' }
+#' }
+#'
+#' @examples
+#' library(flexFitR)
+#' plot_fn(
+#'   fn = "fn_quad_plat",
+#'   params = c(t1 = 35, t2 = 80, b = 4, k = 100),
+#'   interval = c(0, 108),
+#'   n_points = 2000,
+#'   auc_label_size = 3
+#' )
+fn_quad_plat <- function(t, t1 = 45, t2 = 80, b = 1, k = 100) {
+  if (t < t1) {
+    y <- 0
+  } else if (t >= t1 && t <= t2) {
+    c <- suppressWarnings((k - b * (t2 - t1)) / (t2 - t1)^2)
+    y <- b * (t - t1) + c * (t - t1)^2
+  } else {
+    y <- k
+  }
+  return(y)
 }
 
 
@@ -595,56 +658,6 @@ create_call <- function(fn = "fn_lin_plat") {
   return(out)
 }
 
-#' @noRd
-fn_logis <- function(t, t0, t1, t2) {
-  t0 / (1 + exp((t1 - t) / t2))
-}
-
-# library(flexFitR)
-# t <- seq(0, 108, 0.1)
-# y_hat <- sapply(
-#   X = t,
-#   FUN = fn_logistic2,
-#   t1 = 38.7, t2 = 62, L = 0.1, k = 0.9
-# )
-# plot(t, y_hat, type = "l")
-# lines(t, y_hat, col = "red")
-# abline(v = c(38.7, 62), lty = 2)
-#' @noRd
-fn_logistic2 <- function(t, t1, t2, L, k) {
-  # L is the maximum plant height
-  # k is the growth rate
-  if (t < t1) {
-    return(0)
-  }
-  return(L / (1 + exp(-k * (t - (t1 + t2) / 2))))
-}
-
-# t <- seq(0, 108, 0.1)
-# y_hat <- sapply(
-#   X = t,
-#   FUN = fn_linexp,
-#   t1 = 38.7, t2 = 62, a = 0.1, b = 2, c = 0.1
-# )
-# plot(t, y_hat, type = "l")
-# lines(t, y_hat, col = "red")
-# abline(v = c(38.7, 62), lty = 2)
-#' @noRd
-fn_linexp <- function(t, t1, t2, a, b, c) {
-  # a, b, c are model parameters
-  if (t < t1) {
-    return(0)
-  }
-  if (t >= t1 && t <= t2) {
-    return(a + b * (t - t1))
-  }
-  if (t > t2) {
-    y2 <- a + b * (t2 - t1)
-    return(y2 * exp(-c * (t - t2)))
-  }
-}
-
-
 #' Print available functions in flexFitR
 #'
 #' @return A vector with available functions
@@ -659,10 +672,12 @@ list_funs <- function() {
     "fn_quad",
     "fn_logistic",
     "fn_lin_plat",
+    "fn_lin_logis",
+    "fn_quad_plat",
     "fn_lin_pl_lin",
     "fn_lin_pl_lin2",
-    "fn_exp1_exp",
-    "fn_exp1_lin",
+    "fn_exp_exp",
+    "fn_exp_lin",
     "fn_exp2_exp",
     "fn_exp2_lin"
   )