+++ NEWS.md
@@ -5,7 +5,8 @@
 * When evaluating several methods in `modeler()`, Jacobian and Hessian are
 computed only for the best method.
 * Now functions are required to be vectorized (faster execution).
-* Renaming `fn_lin_plat()` function and adding `fn_lin_logis()` and `fn_quad_plat()`.
+* Renaming `fn_lin_plat()` function and adding `fn_lin_logis()`, `fn_quad_plat()` and 
+`fn_quad_pl_sm()`.
 * The `modeler()` function now uses `optimr` instead of `opm` for faster execution.
 * `plot.modeler()` includes `linewidth` argument to increase size in geom lines.
 * `predict.modeler()` includes `parallel` and `workers` to allow for parallel computing.